Backend/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
mypy_extensions==1.1.0
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
